module5-task1 вынес состояние в компоненты
