+++ src/index.tsx
@@ -2,8 +2,6 @@ import React from 'react';
 import ReactDOM from 'react-dom/client';
 import App from './app';
 import {offers} from './mocks/offers';
-import {cities} from './mocks/cities';
-import {comments} from './mocks/comments';
 
 const root = ReactDOM.createRoot(
   document.getElementById('root') as HTMLElement
@@ -11,6 +9,6 @@ const root = ReactDOM.createRoot(
 
 root.render(
   <React.StrictMode>
-    <App offers={offers} cities={cities} comments={comments}/>
+    <App offers={offers}/>
   </React.StrictMode>
 );
